bump a few deps
In particular, allow/require latest major of these in `Cargo.toml`:

* `base64`
* `criterion` (dev only)
* `h264-reader`
* `pretty-hex`
* `rtsp-types`
* `webrtc` (example only)

...and since that pulls in stuff with a minimum Rust of 1.67, bump to
that. Strictly speaking I think it's only dev that requires that now,
but it's old enough that requiring it should be no problem.

I held off on upgrading `bitstream-io` to 2, because the latest
`h264-reader` still uses v1, and I like to minimize how much duplication
I require.