interpret Annex B sequences
Servers are not supposed to use Annex B sequences in RTP payloads, but
V380 cameras do. Retina promises that it "works around brokenness in
cheap closed-source cameras", so let's try to make this work.

As noted by @Curid in #108, there's prior art: gortsp does Annex B
decoding. ffmpeg also appears to support this, perhaps by accident: the
output from its RTP layer is in Annex B format, and it doesn't error on
Annex B sequences in RTP, so callers implicitly get the same result with
proper or improper framing.

Along the way, notice NALs that illegally contain the sequences
`00 00 00` or `00 00 02` and ones that illegally end with `00`. We'll
see if this happens in practice, but I'd prefer to explicitly consider
how to work around it if so, rather than passing along bad data
unknowingly.

Throughput on my laptop falls from 3.4 GiB/s to 2.9 GiB/s on the
`depacketize/h264_aac_discard` test, which uses realistic data.
I think that's acceptable. The code avoids unnecessary copies (or
refcounts even) and uses `memchr::memchr` to optimize the case of
looking for the next zero byte. From skimming a profile, most of the
added time is in `memchr`, so I don't think we're going to do much
better than that.

Fixes #68
Fixes #108