fix RTCP firewall hole punch pkt format
I happened to notice that Wireshark complained about this packet.

This was using a RTCP SR but omitting the mandatory sender info block.
Just switch to using an RR which does not include this block and test
that it validates via our own logic.

Reassuringly, ffmpeg also uses an RR in its `ff_rtp_send_punch_packets`.