maintain `Depacketizer` invariants on `Err`
...in both H.264 and H.265 depacketizers, so that callers can try to
proceed and recover the stream if they so choose, without causing a
panic.

Specifically for the "timestamp changed ... in the middle of a
fragmented NAL" error, enqueue the error but fully process the current
packet, as it may be properly framed (but something before it was
missing).

Also unify the internal interface a bit behind the various
depacketizers.

Fixes #122